requirements.txt
langchain                                0.3.20
langchain-ollama                         0.2.3
ollama                                   0.4.7
pdf2image                                1.17.0
pillow                                   11.1.0